Refactor Swiper components and slides for cleaner code
The commit simplifies and improves the styling of the Swiper components across the application, including the GeneralSwiper and slides in the 'Articles', 'Tips', and 'Courses' sections. The code within these files has been refactored to enhance readability and maintainability. The use of Swiper has been streamlined, resulting in cleaner, more efficient code.