Refactor swiper component for simplicity
The refactor removes the manual registration process for Swiper, replacing it with the use of already developed modules. This eliminates the need for referencing, useEffect hook and global namespace JSX declaration, achieving a more straightforward and less error-prone approach. The new components extend the base props and return the corresponding base components.